Update test to just validate password history count
First attempt to test CIS on 1803 and later used a baseline snapshot of all CIS
policies and registry settings of an unhardened vSphere stemcell and by
adding/updating the password history count to 24.  However, this
baseline was not compatible with GCP stemcell as they seem to have
slightly different baseline values.

As such, the test was simplified to just validate the password history
count and remove all other CIS values.

In addition, updated tests to handle 1803 stemcells as the
`Get-OSVersion` version now returns `windows1803` for 1803 stemcells.

[#165800663](https://www.pivotaltracker.com/story/show/165800663)
 Ensure 'Enforce password history' is set to '24 or more password(s)' 1.1.1 (L1)